// src/lib/constants/enums/errorMessage.ts
export enum ErrorMessage {
  REQUIRED_ERROR = 'Dit veld is verplicht.',
  EMAIL_INVALID = 'Gelieve een geldig emailadres te kiezen.',
  PHONE_INVALID = 'Gelieve een geldig telefoonnummer te kiezen.',
  INVALID_OPTION = 'Gelieve een geldige optie te kiezen.',
  UNSUPPORTED_FILE_FORMAT_ERROR = 'Niet-ondersteund bestandsformaat. Geldige bestanden zijn: png, jpg, jpeg, gif, pdf.',
  MAX_FILE_SIZE_ERROR = 'Het is niet mogelijk bestanden groter dan 5MB te uploaden.',
}